requirements.txt
tensorflow
pandas
matplotlib
requests
pyyaml
